requirements.txt
psycopg2==2.9.6
SQLAlchemy==2.0.3
alembic==1.11.1
python-dotenv==1.0.0